// packages/study-echarts/src/echarts/util/types.ts
export type ColorString = string;
export type ZRColor = ColorString | LinearGradientObject | RadialGradientObject | PatternObject;
// ...
export interface DataHost {
  getData(dataType?: SeriesDataType): SeriesData;
}
// ...
export type StageHandlerProgressParams = TaskProgressParams;
// ...
export type OrdinalRawValue = string | number;
export type OrdinalNumber = number; // The number mapped from each OrdinalRawValue.
// ...
export type ParsedValueNumeric = number | OrdinalNumber;

/**
* `ScaleDataValue` means that the user input primitive value to `src/scale/Scale`.
* (For example, used in `axis.min`, `axis.max`, `convertToPixel`).
* Note:
* `ScaleDataValue` is a little different from `OptionDataValue`, because it will not go through
* `src/data/helper/dataValueHelper.parseDataValue`, but go through `src/scale/Scale['parse']`.
*/
export type ScaleDataValue = ParsedValueNumeric | OrdinalRawValue | Date;
// ...
export type DimensionIndexLoose = DimensionIndex | string;
export type DimensionName = string;
export type DimensionLoose = DimensionName | DimensionIndexLoose;
// ...
export type DimensionDefinition = {
  type?: DataStoreDimensionType,
  name?: DimensionName,
  displayName?: string
};
export type DimensionDefinitionLoose = DimensionDefinition['name'] | DimensionDefinition;
// ...
export const SERIES_LAYOUT_BY_COLUMN = 'column' as const;
export const SERIES_LAYOUT_BY_ROW = 'row' as const;

export type SeriesLayoutBy = typeof SERIES_LAYOUT_BY_COLUMN | typeof SERIES_LAYOUT_BY_ROW;
// null/undefined/'auto': auto detect header, see "src/data/helper/sourceHelper".
// If number, means header lines count, or say, `startIndex`.
// Like `sourceHeader: 2`, means line 0 and line 1 are header, data start from line 2.
export type OptionSourceHeader = boolean | 'auto' | number;

export type SeriesDataType = 'main' | 'node' | 'edge';
// ...
export type OptionDataItemObject<T> = {
  id?: OptionId;
  name?: OptionName;
  groupId?: OptionId;
  value?: T[] | T;
  selected?: boolean;
};
// Compat number because it is usually used and not easy to
// restrict it in practise.
export type OptionId = string | number;
export type OptionName = string | number;
// ...
export interface OptionEncodeVisualDimensions {
  tooltip?: OptionEncodeValue;
  label?: OptionEncodeValue;
  itemName?: OptionEncodeValue;
  itemId?: OptionEncodeValue;
  seriesName?: OptionEncodeValue;
  // Notice: `value` is coordDim, not nonCoordDim.

  // Group id is used for linking the aggregate relationship between two set of data.
  // Which is useful in prepresenting the transition key of drilldown/up animation.
  // Or hover linking.
  itemGroupId?: OptionEncodeValue;
}
export interface OptionEncode extends OptionEncodeVisualDimensions {
  [coordDim: string]: OptionEncodeValue | undefined
}
export type OptionEncodeValue = DimensionLoose | DimensionLoose[];
// ...
export interface BoxLayoutOptionMixin {
  width?: number | string;
  height?: number | string;
  top?: number | string;
  right?: number | string;
  bottom?: number | string;
  left?: number | string;
}
// ...
export type ColorBy = 'series' | 'data';
// ...
export type BlurScope = 'coordinateSystem' | 'series' | 'global';

/**
* can be array of data indices.
* Or may be an dictionary if have different types of data like in graph.
*/
export type InnerFocus = DefaultEmphasisFocus | ArrayLike<number> | Dictionary<ArrayLike<number>>;

export interface DefaultStatesMixin {
  // FIXME
  emphasis?: any
  select?: any
  blur?: any
}

export type DefaultEmphasisFocus = 'none' | 'self' | 'series';

export interface DefaultStatesMixinEmphasis {
  /**
   * self: Focus self and blur all others.
   * series: Focus series and blur all other series.
   */
  focus?: DefaultEmphasisFocus
}

export interface StatesMixinBase {
  emphasis?: unknown
  select?: unknown
  blur?: unknown
}

export interface StatesOptionMixin<
  StateOption,
  StatesMixin extends StatesMixinBase
  > {
  /**
   * Emphasis states
   */
  emphasis?: StateOption & StatesMixin['emphasis'] & {
    /**
     * Scope of blurred element when focus.
     *
     * coordinateSystem: blur others in the same coordinateSystem
     * series: blur others in the same series
     * global: blur all others
     *
     * Default to be coordinate system.
     */
    blurScope?: BlurScope

    /**
     * If emphasis state is disabled.
     */
    disabled?: boolean
  }
  /**
   * Select states
   */
  select?: StateOption & StatesMixin['select'] & {
    disabled?: boolean
  }
  /**
   * Blur states.
   */
  blur?: StateOption & StatesMixin['blur']
}
// ...
export interface SeriesOption<
  StateOption = unknown,
  StatesMixin extends StatesMixinBase = DefaultStatesMixin
  > extends
// ...
export interface SeriesEncodeOptionMixin {
  datasetIndex?: number;
  datasetId?: string | number;
  seriesLayoutBy?: SeriesLayoutBy;
  sourceHeader?: OptionSourceHeader;
  dimensions?: DimensionDefinitionLoose[];
  encode?: OptionEncode
}